:sparkles:Logica de insertar productos iniciales y supermercados añadida

+++ src/main/java/madstodolist/repository/mensajeRepository.java
@@ -4,5 +4,5 @@ import madstodolist.model.Dieta;
 import madstodolist.model.Mensaje;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface mensajeRepository extends JpaRepository<Mensaje, Integer> {
+public interface mensajeRepository extends JpaRepository<Mensaje, Long> {
 }
